황승민 : 엔진 업데이트, 몬스터 이동경로 액터 생성, 컨텐츠 구조 수정

+++ DirectX_HSM_KingdomRushFrontiers/GameEngineContents/MapEdit.cpp
@@ -252,7 +252,7 @@ void MapEdit::SaveData()
 {
 	for (auto& i : Desc.MonsterPaths)
 	{
-		SaveMapData.Write(i.second.size());
+		SaveMapData.Write(static_cast<const int>(i.second.size()));
 		for (auto& j : i.second)
 		{
 			SaveMapData.Write(&j.Position, sizeof(float4));
